Refactor rendering to more reflect schematic files
To more reflect how components are stored in Kicad Schematic files,
new classes were defined to represent Kicad library component symbols.

In a Kicad Schematic file, Kicad stores a token that lists all the types
of symbols (as well as their definitions) that were used in a Schematic.
The rest of the file contains a list of all the instances of these
symbols in the schematic. Two classes were added to reflect this called
SymbolDefinition and SymbolInstance.

A new file was added to mock parsing a schematic file.